pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtpd>=1.4.6",
    "aiosmtplib>=5.1.0",
    "fastapi>=0.135.3",
    "fastapi-mail>=1.6.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
